test: 도메인 테스트 보강
